+++ testing.R
@@ -2,10 +2,9 @@ library(rsimsum)
 library(ggplot2)
 data("tt", package = "rsimsum")
 s6 <- simsum(data = tt, estvarname = "diff", se = "se", true = -1, x = TRUE)
-autoplot(s6, type = "est_ridge")
-
-
-
+autoplot(s6, type = "lolly")
+autoplot(s6, type = "forest")
+autoplot(s6, type = "heat")
 
 ### All-in-one testing
 devtools::document()